Store user input into firebase
