+++ app/src/main/AndroidManifest.xml
@@ -2,36 +2,33 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools">
 
-    <uses-permission android:name="android.permission.INTERNET"/>
+    <uses-permission android:name="android.permission.INTERNET" />
+    <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE"/>
 
     <application
         android:allowBackup="true"
         android:dataExtractionRules="@xml/data_extraction_rules"
         android:fullBackupContent="@xml/backup_rules"
         android:icon="@mipmap/ic_launcher"
-        android:label="Giftly"
+        android:label=""
         android:supportsRtl="true"
         android:theme="@style/Theme.Giftly"
         tools:targetApi="31">
         <activity
-            android:name="HomeScreen"
+            android:name=".HomeScreen"
             android:exported="false" />
         <activity
-            android:name="LoginScreen"
+            android:name=".SignUpScreen"
             android:exported="false" />
         <activity
-            android:name="SignUpScreen"
-            android:exported="false" />
-
-        <activity
-            android:name="MainActivity"
+            android:name=".MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
+
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-
     </application>
 
 </manifest>
