+++ app/src/main/java/com/example/giftly/HomeScreen.java
@@ -6,7 +6,6 @@ import android.os.Bundle;
 
 public class HomeScreen extends AppCompatActivity {
 
-    //Variables for profile icon
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
