--- app/src/main/res/layout/activity_login_screen.xml
@@ -1,83 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:background="#1C1C1C"
-    tools:context=".LoginScreen">
-
-    <!-- User name entry option -->
-
-    <ImageView
-        android:id="@+id/imageView"
-        android:layout_width="172dp"
-        android:layout_height="137dp"
-
-        android:layout_marginStart="120dp"
-        android:layout_marginTop="85dp"
-        android:layout_marginEnd="119dp"
-        android:layout_marginBottom="17dp"
-        app:layout_constraintBottom_toTopOf="@+id/usernm1"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:srcCompat="@drawable/giftly_icon" />
-
-    <!-- Password entry option -->
-
-
-    <!-- Sign in button -->
-
-    <EditText
-        android:id="@+id/usernm1"
-        android:layout_width="309dp"
-        android:layout_height="78dp"
-        android:layout_marginStart="51dp"
-        android:layout_marginTop="17dp"
-        android:layout_marginEnd="51dp"
-        android:layout_marginBottom="2dp"
-        android:fontFamily="sans-serif-medium"
-        android:hint="User Name"
-        android:textColor="#ffffff"
-        android:textColorHint="#ffffff"
-        app:layout_constraintBottom_toTopOf="@+id/userpasswrd1"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/imageView" />
-
-    <EditText
-        android:id="@+id/userpasswrd1"
-        android:layout_width="309dp"
-        android:layout_height="78dp"
-        android:layout_marginStart="51dp"
-        android:layout_marginTop="1dp"
-        android:layout_marginEnd="51dp"
-        android:layout_marginBottom="114dp"
-        android:fontFamily="sans-serif-medium"
-        android:hint="Password"
-        android:textColor="#ffffff"
-        android:textColorHint="#ffffff"
-        app:layout_constraintBottom_toTopOf="@+id/signsbutton1"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/usernm1" />
-
-    <Button
-        android:id="@+id/signsbutton1"
-        android:layout_width="180dp"
-        android:layout_height="65dp"
-        android:layout_marginStart="119dp"
-        android:layout_marginTop="50dp"
-        android:layout_marginEnd="112dp"
-        android:layout_marginBottom="133dp"
-        android:text="Log In"
-        app:backgroundTint="#008CF8"
-        app:cornerRadius="40dp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/userpasswrd1" />
-
-
-</androidx.constraintlayout.widget.ConstraintLayout>
