+++ app/src/main/res/layout/activity_sign_up_screen.xml
@@ -26,7 +26,7 @@
         android:layout_marginTop="-10dp"
         android:layout_marginEnd="37dp"
         android:fontFamily="sans-serif-medium"
-        android:hint="User Name"
+        android:hint="Name"
         android:textColor="#ffffff"
         android:textColorHint="#ffffff"
         app:layout_constraintEnd_toEndOf="parent"
@@ -64,6 +64,21 @@
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/editEmail" />
 
+    <EditText
+        android:id="@+id/userInterests"
+        android:layout_width="309dp"
+        android:layout_height="78dp"
+        android:layout_marginStart="66dp"
+        android:layout_marginTop="-17dp"
+        android:layout_marginEnd="36dp"
+        android:fontFamily="sans-serif-medium"
+        android:hint="Enter Your Interests"
+        android:textColor="#ffffff"
+        android:textColorHint="#ffffff"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/editPassword" />
+
     <Button
         android:id="@+id/btnSubmit"
         android:layout_width="197dp"
@@ -75,14 +90,13 @@
         app:cornerRadius="40dp"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/editPassword" />
+        app:layout_constraintTop_toBottomOf="@+id/userInterests" />
 
     <TextView
         android:id="@+id/txtLoginInfo"
         android:layout_width="205dp"
         android:layout_height="33dp"
         android:layout_marginStart="103dp"
-        android:layout_marginTop="9dp"
         android:layout_marginEnd="103dp"
         android:fontFamily="sans-serif-medium"
         android:text="Already have an account? Log in"
